drivers: eth: e1000: Use double with PTP clock
Instead of mixing floats and doubles, convert the code to
use double so that we avoid float->double conversion warning
from compiler.

+++ drivers/ethernet/eth_e1000_priv.h
@@ -92,7 +92,7 @@ struct e1000_dev {
 	uint8_t rxb[NET_ETH_MTU];
 #if defined(CONFIG_ETH_E1000_PTP_CLOCK)
 	const struct device *ptp_clock;
-	float clk_ratio;
+	double clk_ratio;
 #endif
 };
